dupla_packages/dupla_discal/requirements.txt
matplotlib==3.4.2
numba==0.53.1
numpy==1.21.0
pandas==1.2.5
Pillow==9.3.0
pycpd==2.0.0
pytest==6.2.4
scikit_image==0.18.2
scikit_learn==1.2.0
scipy==1.7.0
